jump to idf 5.3

+++ components/rgbHsv/include/rgbHsv.h
@@ -25,4 +25,4 @@ HsvColor RgbToHsv(RgbColor rgb);
 void parseRGB(RgbColor *color, char* payload);
 uint8_t modeToEnum(char* str);
 
-uint8_t checkColorAndBright(RgbColor *currentRGB, RgbColor *targetRGB, uint8_t *currentBright, uint8_t *targetBright, uint8_t fade_increment);
+uint8_t checkColorAndBright(RgbColor *currentRGB, RgbColor *targetRGB, int16_t *currentBright, int16_t *targetBright, int16_t fade_increment);
